+++ level/errors.go
@@ -1,0 +1,9 @@
+// Copyright 2021 The sypl Authors. All rights reserved.
+// Use of this source code is governed by a MIT
+// license that can be found in the LICENSE file.
+
+package level
+
+import "errors"
+
+var ErrInvalidLevel = errors.New("invalid error level")