+++ output/output.go
@@ -170,9 +170,9 @@ func (o *output) SetWriter(w io.Writer) {
 	o.writer = w
 }
 
-// Write write the message to the defined output.
-//
-// TODO: Review complexity.
+// Write the message to the defined output. In case of any error, it can be
+// introspected, providing more information about the failure. The error will be
+// the type of `ProcessingError`.
 //nolint:nestif
 func (o *output) Write(m message.IMessage) error {
 	// Should allows to specify `Output`(s).